KS-112 inittest for salt in tests
