+++ tests/test_basic_connectivity.py
@@ -10,17 +10,12 @@ Test Basic Connectivity
 from __future__ import (
     absolute_import, unicode_literals, division, print_function)
 import unittest
-import uuid
 from kount.response import Response
-from kount.request import (ASTAT, BCRSTAT, INQUIRYMODE,
-                           CURRENCYTYPE, MERCHANTACKNOWLEDGMENT)
-from kount.inquiry import Inquiry
-from kount.util.payment import CardPayment
-from kount.util.cartitem import CartItem
-from kount.util.address import Address
 from kount.util.ris_validation_exception import RisValidationException
 from kount.client import Client
-from kount.settings import SDK_VERSION, RAISE_ERRORS
+from kount.settings import RAISE_ERRORS
+import inittest
+from test_inquiry import generate_unique_id, default_inquiry
 
 __author__ = "Yordanka Spahieva"
 __version__ = "1.0.0"
@@ -29,10 +24,6 @@ __email__ = "yordanka.spahieva@sirma.bg"
 __status__ = "Development"
 
 
-BILLING_ADDRESS = Address("1234 North B2A1 Tree Lane South",
-                          "", "Albuquerque", "NM", "87101", "US")
-SHIPPING_ADDRESS = Address("567 West S2A1 Court North", "",
-                           "Gnome", "AK", "99762", "US")
 URL_API = "https://risk.beta.kount.net"
 URL_API_BETA = URL_API
 MERCHANT_ID6 = '999666'
@@ -40,50 +31,6 @@ MERCHANT_ID7 = '999667'
 PTOK = "0007380568572514"
 KOUNT_API_KEY6 = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiI5OTk2NjYiLCJhdWQiOiJLb3VudC4xIiwiaWF0IjoxNDk0NTM0Nzk5LCJzY3AiOnsia2EiOm51bGwsImtjIjpudWxsLCJhcGkiOmZhbHNlLCJyaXMiOnRydWV9fQ.eMmumYFpIF-d1up_mfxA5_VXBI41NSrNVe9CyhBUGck"
 KOUNT_API_KEY7 = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiI5OTk2NjciLCJhdWQiOiJLb3VudC4xIiwiaWF0IjoxNDk0NTM1OTE2LCJzY3AiOnsia2EiOm51bGwsImtjIjpudWxsLCJhcGkiOmZhbHNlLCJyaXMiOnRydWV9fQ.KK3zG4dMIhTIaE5SeCbej1OAFhZifyBswMPyYFAVRrM"
-
-
-def generate_unique_id():
-    "unique session id"
-    return str(uuid.uuid4()).replace('-', '').upper()
-
-
-def default_inquiry(session_id, merchant_id, email_client, ptok):
-    "default_inquiry, PENC is not set"
-    result = Inquiry()
-    result.request_mode(INQUIRYMODE.DEFAULT)
-    result.shipping_address(SHIPPING_ADDRESS)
-    result.shipping_name("SdkShipToFN SdkShipToLN") #S2NM
-    result.billing_address(BILLING_ADDRESS)
-    result.currency_set(CURRENCYTYPE.USD)   #CURR
-    result.total_set('123456') #TOTL
-    result.billing_phone_number("555-867-5309") #B2PN
-    result.shipping_phone_number("555-777-1212") #S2PN
-    result.email_client(email_client)
-    result.customer_name("SdkTestFirstName SdkTestLastName")
-    result.unique_customer_id(session_id[:20]) #UNIQ
-    result.website("DEFAULT") #SITE
-    result.email_shipping("sdkTestShipToEmail@kountsdktestdomain.com")
-    result.ip_address("4.127.51.215") #IPAD
-    cart_item = []
-    cart_item.append(CartItem("SPORTING_GOODS", "SG999999",
-                              "3000 CANDLEPOWER PLASMA FLASHLIGHT",
-                              '2', '68990'))
-    result.shopping_cart(cart_item)
-    result.version()
-    result.version_set(SDK_VERSION)  #0695
-    result.merchant_set(merchant_id)
-    payment = CardPayment(ptok)
-    result.payment_set(payment) #PTOK
-    result.session_set(session_id) #SESS
-    result.order_number(session_id[:10])  #ORDR
-    result.authorization_status(ASTAT.Approve) #AUTH
-    result.avs_zip_reply(BCRSTAT.MATCH)
-    result.avs_address_reply(BCRSTAT.MATCH)
-    result.avs_cvv_reply(BCRSTAT.MATCH)
-    result.merchant_acknowledgment_set(MERCHANTACKNOWLEDGMENT.TRUE) #"MACK"
-    result.cash('4444')
-    #~ result.params["PENC"] = "KHASH"
-    return result
 
 
 class TestBasicConnectivity(unittest.TestCase):
@@ -133,7 +80,7 @@ class TestBasicConnectivity(unittest.TestCase):
         self.inq.params["S2NM"] = bad
         self.inq.params["EMAL"] = bad
         self.assertRaises(
-            ValueError,
+            RisValidationException,
             Client(URL_API, KOUNT_API_KEY7,
                    raise_errors=True).process, self.inq.params)
         res = Client(URL_API, KOUNT_API_KEY7,
@@ -169,7 +116,7 @@ class TestBasicConnectivity(unittest.TestCase):
         for bad in bad_list:
             inq.params["S2NM"] = bad * 999
             self.assertRaises(
-                ValueError,
+                RisValidationException,
                 Client(URL_API,
                        KOUNT_API_KEY6,
                        raise_errors=True).process, inq.params)
