+++ tests/test_inquiry.py
@@ -7,8 +7,14 @@
 from __future__ import (
     absolute_import, unicode_literals, division, print_function)
 import unittest
+import uuid
+from kount.request import (ASTAT, BCRSTAT, INQUIRYMODE,
+                           CURRENCYTYPE, MERCHANTACKNOWLEDGMENT)
+from kount.inquiry import Inquiry
+from kount.util.payment import CardPayment
+from kount.util.cartitem import CartItem
+from kount.util.address import Address
 from kount.settings import SDK_VERSION
-from test_basic_connectivity import generate_unique_id, default_inquiry
 
 __author__ = "Yordanka Spahieva"
 __version__ = "1.0.0"
@@ -20,6 +26,56 @@ __status__ = "Development"
 EMAIL_CLIENT = "sdkTest@kountsdktestdomain.com"
 MERCHANT_ID = '999666'
 PTOK = "0007380568572514"
+
+
+BILLING_ADDRESS = Address("1234 North B2A1 Tree Lane South",
+                          "", "Albuquerque", "NM", "87101", "US")
+SHIPPING_ADDRESS = Address("567 West S2A1 Court North", "",
+                           "Gnome", "AK", "99762", "US")
+
+
+def generate_unique_id():
+    "unique session id"
+    return str(uuid.uuid4()).replace('-', '').upper()
+
+
+def default_inquiry(session_id, merchant_id, email_client, ptok):
+    "default_inquiry, PENC is not set"
+    result = Inquiry()
+    result.request_mode(INQUIRYMODE.DEFAULT)
+    result.shipping_address(SHIPPING_ADDRESS)
+    result.shipping_name("SdkShipToFN SdkShipToLN") #S2NM
+    result.billing_address(BILLING_ADDRESS)
+    result.currency_set(CURRENCYTYPE.USD)   #CURR
+    result.total_set('123456') #TOTL
+    result.billing_phone_number("555-867-5309") #B2PN
+    result.shipping_phone_number("555-777-1212") #S2PN
+    result.email_client(email_client)
+    result.customer_name("SdkTestFirstName SdkTestLastName")
+    result.unique_customer_id(session_id[:20]) #UNIQ
+    result.website("DEFAULT") #SITE
+    result.email_shipping("sdkTestShipToEmail@kountsdktestdomain.com")
+    result.ip_address("4.127.51.215") #IPAD
+    cart_item = []
+    cart_item.append(CartItem("SPORTING_GOODS", "SG999999",
+                              "3000 CANDLEPOWER PLASMA FLASHLIGHT",
+                              '2', '68990'))
+    result.shopping_cart(cart_item)
+    result.version()
+    result.version_set(SDK_VERSION)  #0695
+    result.merchant_set(merchant_id)
+    payment = CardPayment(ptok)
+    result.payment_set(payment) #PTOK
+    result.session_set(session_id) #SESS
+    result.order_number(session_id[:10])  #ORDR
+    result.authorization_status(ASTAT.Approve) #AUTH
+    result.avs_zip_reply(BCRSTAT.MATCH)
+    result.avs_address_reply(BCRSTAT.MATCH)
+    result.avs_cvv_reply(BCRSTAT.MATCH)
+    result.merchant_acknowledgment_set(MERCHANTACKNOWLEDGMENT.TRUE) #"MACK"
+    result.cash('4444')
+    #~ result.params["PENC"] = "KHASH"
+    return result
 
 
 class TestInquiry(unittest.TestCase):
